requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Streamlit UI
streamlit==1.31.0
streamlit-authenticator==0.3.1
streamlit-chat==0.1.1

# OpenAI & AI
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20

# Azure Services
azure-ai-formrecognizer==3.3.2
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# Document Processing
pypdf==4.0.1
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.2.0
pillow==10.2.0
pdf2image==1.17.0
pytesseract==0.3.10

# OCR & Image Processing
opencv-python==4.9.0.80
numpy==1.26.3

# Text Processing & NLP
spacy==3.7.2
nltk==3.8.1
langdetect==1.0.9
tiktoken==0.5.2

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

